requirements.txt
loguru==0.7.3
awscli==1.37.14
pandas==2.2.3
pyarrow==20.0.0
Pillow==9.4.0
polars==1.31.0
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0